requirements.txt
matplotlib==3.9.2
pandas==2.2.3
numpy==2.1.3
pycoingecko==3.2.0
plotly==5.24.1
mplfinance==0.12.10b0
requests==2.32.3
pillow==11.0.0
